+++ apps/web/modules/event-types/components/__tests__/AddMembersWithSwitch.test.tsx
@@ -14,7 +14,7 @@ vi.mock("@formkit/auto-animate/react", () => ({
 }));
 
 // Mock Segment component
-vi.mock("@calcom/features/Segment", () => ({
+vi.mock("@calcom/web/modules/event-types/components/Segment", () => ({
   Segment: vi.fn().mockImplementation(({ onQueryValueChange }) => (
     <div data-testid="mock-segment">
       <button onClick={() => onQueryValueChange({ queryValue: { id: "test" } })}>Update Query</button>
