+++ apps/web/modules/settings/license-key/new/new-view.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { CreateANewLicenseKeyForm } from "@calcom/features/ee/deployment/licensekey/CreateLicenseKeyForm";
+import { CreateANewLicenseKeyForm } from "@calcom/web/modules/ee/deployment/licensekey/CreateLicenseKeyForm";
 import { WizardLayout } from "@calcom/ui/components/layout";
 
 export const LayoutWrapper = ({ children }: { children: React.ReactNode }) => {
